+++ src/components/pronostics/pronostics.css
@@ -1,0 +1,3 @@
+.p-tag {
+  color: blue;
+}
