--- src/pages/pronostics/pronostic.css
@@ -1,7 +1,0 @@
-.content {
-  background-image: linear-gradient(115deg, #e60a0aaf, #163ff1d3) url('C:\Users\HP\Desktop\rebasepractice\fakescores\public\assets\images\scorematch.png');
-}
-
-h1 {
-  color: red;
-}
